pyproject.toml
[tool.poetry]
name = "assessment_due_dates"
version = "0.1.0"
description = ""
authors = ["Peter Marsh <p.marsh@mac.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
